Remove 'Home Directory'.#214
