+++ jp.go.aist.rtm.rtcbuilder.nl1/src/jp/go/aist/rtm/rtcbuilder/nl/messages_ja.properties
@@ -480,10 +480,6 @@ IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_TITLE=\u30d7\u30ed\u30b8\u30a7\u3
 IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_P1=\u6307\u5b9a\u3055\u308c\u305f\u30d7\u30ed\u30b8\u30a7\u30af\u30c8\u304cWorkspace\u5185\u306b\u5b58\u5728\u3057\u307e\u305b\u3093\u3002
 IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_P2=\u65b0\u898f\u306b\u751f\u6210\u3057\u3066\u3082\u3088\u308d\u3057\u3044\u3067\u3059\u304b\uff1f
 
-IPreferenceMessageConstants.HOME_DIR=\u30db\u30fc\u30e0\u30c7\u30a3\u30ec\u30af\u30c8\u30ea
-IPreferenceMessageConstants.DATA_TYPE_BTN_ADD=Add
-IPreferenceMessageConstants.DATA_TYPE_BTN_DELETE=Delete
-
 IPreferenceMessageConstants.PREFIX=\u63a5\u982d\u8a9e
 IPreferenceMessageConstants.SUFFIX=\u63a5\u5c3e\u8a9e
