+++ jp.go.aist.rtm.rtcbuilder/plugin.properties
@@ -37,8 +37,8 @@ perspectiveName = RTC Builder
 preferecePagesName = RtcBuilder
 preferecePagesCodeGenerateName = Code Generate
 preferecePagesPortsName = Port
-preferecePagesDocumentName = Document
-preferecePagesDataTypeName = Data Type
+preferecePagesDocumentName = Activity, Document
+preferecePagesConfigurationName = Configuration
 
 #Content Types
 rtcProfileName = RTC Profile
