+++ jp.go.aist.rtm.rtcbuilder/plugin.xml
@@ -109,6 +109,11 @@
            class="jp.go.aist.rtm.rtcbuilder.ui.preference.DocumentPreferencePage"
            id="jp.go.aist.rtm.rtcbuilder.ui.preferencePage.codeGenerate.Document"
            name="%preferecePagesDocumentName"/>
+     <page
+           category="jp.go.aist.rtm.rtcbuilder.ui.preferencePage.codeGenerate"
+           class="jp.go.aist.rtm.rtcbuilder.ui.preference.CodeGenerateConfigPreferencePage"
+           id="jp.go.aist.rtm.rtcbuilder.ui.preferencePage.codeGenerate.Configuration"
+           name="%preferecePagesConfigurationName"/>
   </extension>
   <extension
         point="org.eclipse.core.runtime.contentTypes">
