+++ jp.go.aist.rtm.rtcbuilder/src/jp/go/aist/rtm/rtcbuilder/nl/messages.properties
@@ -481,10 +481,6 @@ IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_TITLE=Project Generation
 IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_P1=The specified project does not exist in Workspace.
 IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_P2=Are you sure you want to generate new project?
 
-IPreferenceMessageConstants.HOME_DIR=Home Directory
-IPreferenceMessageConstants.DATA_TYPE_BTN_ADD=Add
-IPreferenceMessageConstants.DATA_TYPE_BTN_DELETE=Delete
-
 IPreferenceMessageConstants.PREFIX=Prefix
 IPreferenceMessageConstants.SUFFIX=Suffix
