+++ jp.go.aist.rtm.rtcbuilder/src/jp/go/aist/rtm/rtcbuilder/ui/preference/CodeGeneratePreferencePage.java
@@ -1,9 +1,5 @@
 package jp.go.aist.rtm.rtcbuilder.ui.preference;
 
-import jp.go.aist.rtm.rtcbuilder.RtcBuilderPlugin;
-import jp.go.aist.rtm.rtcbuilder.nl.Messages;
-import jp.go.aist.rtm.rtcbuilder.ui.editors.IMessageConstants;
-
 import org.eclipse.jface.preference.ComboFieldEditor;
 import org.eclipse.jface.preference.IPreferenceStore;
 import org.eclipse.jface.preference.IntegerFieldEditor;
@@ -14,6 +10,9 @@ import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.ui.IWorkbench;
 import org.eclipse.ui.IWorkbenchPreferencePage;
+
+import jp.go.aist.rtm.rtcbuilder.RtcBuilderPlugin;
+import jp.go.aist.rtm.rtcbuilder.nl.Messages;
 
 public class CodeGeneratePreferencePage extends AbstarctFieldEditorPreferencePage implements
 							IWorkbenchPreferencePage {
@@ -46,8 +45,6 @@ public class CodeGeneratePreferencePage extends AbstarctFieldEditorPreferencePag
 	public void init(IWorkbench workbench) {
 		IPreferenceStore store = RtcBuilderPlugin.getDefault().getPreferenceStore();
 		storeComponentInitialSetting(store);
-		storeConfigurationSetInitialSetting(store);
-		storeBackupInitialSetting(store);
 	}
 
 	@Override
@@ -58,52 +55,6 @@ public class CodeGeneratePreferencePage extends AbstarctFieldEditorPreferencePag
 		gd.grabExcessHorizontalSpace = true;
 		composite.setLayoutData(gd);
 		createComponentPart(composite);
-		createConfigurationSetParts(composite);
-		createBackupParts(composite);
-	}
-
-	private void createBackupParts(Composite composite) {
-		Composite backupGroup = createGroup(composite, IPreferenceMessageConstants.CODE_GEN_TITLE_BACKUP);
-		IntegerFieldEditor moduleMaxInstanceTextEditor = new IntegerFieldEditor(ComponentPreferenceManager.Generate_Backup_Num,
-				IMessageConstants.BACKUP_FILE_NUM, backupGroup);
-		addField(moduleMaxInstanceTextEditor);
-	}
-	
-	private void createConfigurationSetParts(Composite composite) {
-		Composite configGroup = createGroup(composite, IPreferenceMessageConstants.CODE_GEN_TITLE_CONFIG);
-		DigitAlphabetStringFieldEditor configurationNameEditor = 
-			new DigitAlphabetStringFieldEditor(ComponentPreferenceManager.Generate_Configuration_Name,
-					Messages.getString("IMC.CONFIGURATION_TBLLBL_NAME"), configGroup);
-		addField(configurationNameEditor);
-		StringFieldEditor configurationTypeEditor = 
-			new StringFieldEditor(ComponentPreferenceManager.Generate_Configuration_Type,
-					Messages.getString("IMC.CONFIGURATION_TBLLBL_TYPE"), configGroup);
-		addField(configurationTypeEditor);
-		StringFieldEditor configurationVarNameEditor = 
-			new StringFieldEditor(ComponentPreferenceManager.Generate_Configuration_VarName,
-					Messages.getString("IMC.CONFIGURATION_LBL_VARNAME"), configGroup);
-		addField(configurationVarNameEditor);
-		DigitAlphabetStringFieldEditor configurationDefaultEditor = 
-			new DigitAlphabetStringFieldEditor(ComponentPreferenceManager.Generate_Configuration_Default,
-					Messages.getString("IMC.CONFIGURATION_TBLLBL_DEFAULTVAL"), configGroup);
-		addField(configurationDefaultEditor);
-		StringFieldEditor configurationConstraintEditor = 
-			new StringFieldEditor(ComponentPreferenceManager.Generate_Configuration_Constraint,
-					Messages.getString("IMC.CONFIGURATION_LBL_CONSTRAINT"), configGroup);
-		addField(configurationConstraintEditor);
-		StringFieldEditor configurationUnitEditor = 
-			new StringFieldEditor(ComponentPreferenceManager.Generate_Configuration_Unit,
-					Messages.getString("IMC.CONFIGURATION_LBL_UNIT"), configGroup);
-		addField(configurationUnitEditor);
-		//
-		StringFieldEditor configurationPrefixEditor = 
-			new StringFieldEditor(ComponentPreferenceManager.Generate_Configuration_Prefix,
-					IPreferenceMessageConstants.PORT_LBL_PREFIX, configGroup);
-		addField(configurationPrefixEditor);
-		StringFieldEditor configurationSuffixEditor = 
-			new StringFieldEditor(ComponentPreferenceManager.Generate_Configuration_Suffix,
-					IPreferenceMessageConstants.PORT_LBL_SUFFIX, configGroup);
-		addField(configurationSuffixEditor);
 	}
 
 	private void createComponentPart(Composite composite) {
@@ -164,21 +115,6 @@ public class CodeGeneratePreferencePage extends AbstarctFieldEditorPreferencePag
 		addField(commonSuffixEditor);
 	}
 	
-	private void storeBackupInitialSetting(IPreferenceStore store) {
-		store.setDefault(ComponentPreferenceManager.Generate_Backup_Num, ComponentPreferenceManager.DEFAULT_BACKUP_NUM);
-	}
-
-	private void storeConfigurationSetInitialSetting(IPreferenceStore store) {
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_Name, ComponentPreferenceManager.DEFAULT_CONFIGURATION_NAME);
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_Type, ComponentPreferenceManager.DEFAULT_CONFIGURATION_TYPE);
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_VarName, ComponentPreferenceManager.DEFAULT_CONFIGURATION_VARNAME);
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_Default, ComponentPreferenceManager.DEFAULT_CONFIGURATION_DEFAULT);
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_Constraint, ComponentPreferenceManager.DEFAULT_CONFIGURATION_CONSTRAINT);
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_Unit, ComponentPreferenceManager.DEFAULT_CONFIGURATION_UNIT);
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_Prefix, ComponentPreferenceManager.DEFAULT_CONFIGURATION_PREFIX);
-		store.setDefault(ComponentPreferenceManager.Generate_Configuration_Suffix, ComponentPreferenceManager.DEFAULT_CONFIGURATION_SUFFIX);
-	}
-
 	private void storeComponentInitialSetting(IPreferenceStore store) {
 		store.setDefault(ComponentPreferenceManager.Generate_Basic_Name, ComponentPreferenceManager.DEFAULT_COMPONENT_NAME);
 		store.setDefault(ComponentPreferenceManager.Generate_Basic_Description, ComponentPreferenceManager.DEFAULT_DESCRIPTION);
