+++ jp.go.aist.rtm.rtcbuilder/src/jp/go/aist/rtm/rtcbuilder/util/RTCUtil.java
@@ -1,6 +1,5 @@
 package jp.go.aist.rtm.rtcbuilder.util;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -12,11 +11,9 @@ import org.eclipse.core.resources.IWorkspaceRoot;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.swt.graphics.RGB;
 
-import jp.go.aist.rtm.rtcbuilder.RtcBuilderPlugin;
 import jp.go.aist.rtm.rtcbuilder.generator.param.DataTypeParam;
 import jp.go.aist.rtm.rtcbuilder.generator.param.RtcParam;
 import jp.go.aist.rtm.rtcbuilder.generator.param.idl.IdlPathParam;
-import jp.go.aist.rtm.rtcbuilder.ui.preference.RTCBuilderPreferenceManager;
 
 public class RTCUtil {
 
@@ -61,34 +58,6 @@ public class RTCUtil {
 		return result;
 	}
 
-	public static void setDefaultUserDir() {
-		if(RtcBuilderPlugin.getDefault()!=null) {
-			String resultTemp = RtcBuilderPlugin.getDefault().getPreferenceStore().getString(RTCBuilderPreferenceManager.HOME_DIRECTORY);
-			if(resultTemp==null || resultTemp.isEmpty()) {
-				boolean isWindows = false;
-				String targetOS = System.getProperty("os.name").toLowerCase();
-				if(targetOS.toLowerCase().startsWith("windows")) {
-					isWindows = true;
-				}
-				String dirName = "";
-				String FS = System.getProperty("file.separator");
-				if(isWindows) {
-					dirName = System.getenv("APPDATA");
-				} else {
-					dirName = System.getProperty("user.home");
-				}
-				String userHome = dirName + FS + ".openrtp";
-				dirName += FS + ".openrtp" + FS + "idl";
-				File newdir = new File(dirName);
-				try {
-					newdir.mkdirs();
-				} catch (Exception ex) {
-				}
-				RtcBuilderPlugin.getDefault().getPreferenceStore().setValue(RTCBuilderPreferenceManager.HOME_DIRECTORY, userHome);
-			}
-		}
-	}
-
 	public static void getIDLPathes(RtcParam target) {
 		target.getIdlSearchPathList().clear();
 		List<String> added = new ArrayList<String>();
@@ -102,16 +71,6 @@ public class RTCUtil {
 			}
 			target.getIdlSearchPathList().add(new IdlPathParam(defaultPath + "rtm" + FS + "idl", true));
 			added.add(defaultPath + "rtm" + FS + "idl");
-		}
-		//
-		if(RtcBuilderPlugin.getDefault()!=null) {
-			RtcBuilderPlugin.getDefault().getPreferenceStore().setDefault(RTCBuilderPreferenceManager.HOME_DIRECTORY, "");
-			String userHome = RtcBuilderPlugin.getDefault().getPreferenceStore().getString(RTCBuilderPreferenceManager.HOME_DIRECTORY);
-			String userDir = userHome + FS + "idl";
-			if(added.contains(userDir)==false) {
-				target.getIdlSearchPathList().add(new IdlPathParam(userDir, false));
-				added.add(userDir);
-			}
 		}
 		//
 		if(target!=null && target.getOutputProject()!=null && 0<target.getOutputProject().length()) {
